sg: Hack around broken TLB setup
The simple guest puts the temporary AS1 mapping in the entry after
the one it was booted with... and then uses TLB1[1] for the permanent mapping.

Since we boot it with TLB1[0], that means the temporary mapping is in TLB1[1].

The proper fix is to do as the kernel does, but for now make it use
TLB1[2] to get things working.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>