Link in libfdt
By default, libfdt is looked for in ../dtc/libfdt; this can be overridden
by 'make LIBFDT_DIR=<dir>'.

All output is now directed into a "bin" directory, rather than placed in the source directory.
This makes it easier to deal with out-of-tree sources.

The old dependency sed macros are removed; they didn't seem to be actually doing anything useful.

memchr() is added, as libfdt depends on it.  memrchr() is added as well, while I'm at it.

Signed-off-by: Scott Wood <scottwood@freescale.com>