Preliminary device tree support
The hypervisor is passed a master device tree from u-boot, which
contains guest partitions as binary includes.  The hypervisor
populates its guest physical to physical page tables based on
the translatable reg properties in the guest device tree.

There's a fair bit of miscellaneous stuff in here; sorry for
not breaking it up.  Let me know if anything overly weird got left in.