Add msgsnd emulation
Add function emu_msgsnd(), which is called when the guest tries to execute
the msgsnd instruction.

Add function guest_critical_doorbell(), which is called when the hypervisor
receives a critical doorbell exception.  Currently, this happens only when
the hypervisor tries to execute msgsnd with type == 3.  This funciton reflects
the critical doorbell execpetion to the guest.

Updated guest_t to store the count of guest->gcpus[].  emu_msgsnd() needs this
to validate the PIR field.

Added test program msgsnd_test.c.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>