Fix GPR values when booting guests
Make sure all registers r3 - r9 have the correct values when booting the guest.
All registers are specified as inputs because this guarantees that they will
contain the right values and that the compiler won't re-use them.  If they
are specified as clobbers only, then the compiler might use them for %0 or %1.

Signed-off-by: Timur Tabi <timur@freescale.com>