Revert "add rootfs target to linux makefile"
This reverts commit 206b972ccb770b83b9fbd2a16868d978303476e4.

We do not want to require sudo access simply to run the Linux test.  This
was discussed and reverted in the past; please do not reintroduce such
things without reopening the discussion.  Also, this patch was not sent
to hvpatch; please do not apply patches that you haven't posted for
discussion.

Furthermore, the rootfs target will not work when the image is on an NFS
partition with root squashing, and the image is not world-writeable.  The
gunzip command should also have the -f flag, so that subsequent builds
aren't interrupted with an "overwrite?" prompt.

Signed-off-by: Scott Wood <scottwood@freescale.com>