Fix ELF entry point handling (Bug 1544)
Fixed a sign error in the relative entry calculation, and made the entry
point word-sized (or else it would load the high register in the asm
constraint).

The entry point can never be over 4G with 32-bit virtual, because we use
an identity mapping.  Target should also probably be word-sized because
of this.

Signed-off-by: Scott Wood <scottwood@freescale.com>