Make ppid-to-liodn processing more efficient & fixing a bug in it's passthru.
Earlier there was a nested loop for ppid-to-liodn search in hypervisor
device tree for each ppid node processing, and ppid-to-liodn search in
guest device tree(s) for passthru of the same property. Also the passthru
was being done for the first node in the guest device tree(s). Now,
the first instance of ppid-to-liodn table is searched and cached and there
are simplified loops for ppid processing and ppid-to-liodn passthru's.

Signed-off-by: Ashish Kalra <ashish.kalra@freescale.com>