Synchronization fixes.
Use lwsync for load/load and load/store and msync for store/load.

Some spin loops where mbar was used have data dependencies, so only
barrier is required.

msgsnd is treated as a store with respect to mbar, so the msync/isync
was overkill.

setevent/setgevent need to synchronize with previous stores, but don't
need an mbar before actually sending the doorbell, because
send*doorbell() includes it.

Remove a duplicate clearing of ret_user_hook.

Signed-off-by: Scott Wood <scottwood@freescale.com>