Fix bcmux locking.
Lock both tx and rx when adding a mux.  Skip the lock on the transmit
path when crashing.

Remove a bogus error message in the transmit path; it's perfectly
possible to have no data in a data_avail callback if there was a race
with the outbound byte channel's space_avail callback (bug 2129).

Also, some minor code formatting cleanups.

Signed-off-by: Scott Wood <scottwood@freescale.com>