hv: add watchdog emulation
Emulate the watchdog timer for guests.  If the watchdog is allowed to expire
twice, it will generate a processor reset, which cannot be allowed.  So when
the watchdog timer would have caused a processor reset, the hypervisor needs
to catch that and restart the partition instead.

The hypervisor keeps track of watchdog timer interrupts.  When one occurs,
the hypervisor remembers.  If another interrupt occurs before the guest
resets the timer, then the hypervisor will restart the partition.

Signed-off-by: Timur Tabi <timur@freescale.com>