hv: add support for fsl,hv-wd-mgr-notify property
Add code which checks for the "fsl,hv-wd-mgr-notify" in the hypervisor device
tree /partitions node, and stores the result in an internal value.  Update the
watchdog trap handler to respect TCR[WRC] and the manager-notify option.
Sending a message to the manager partition is not currently supported,
since the necessary infrastructure has not yet been implemented.

Also update the watchdog unit test to check for TCR[WRC]==0.

Signed-off-by: Timur Tabi <timur@freescale.com>