hv: define state change, watchdog, and reset request doorbells
Define doorbell nodes for managed partitions to signal the managers for special
events: stat change, watchdog expiration, and reset request.  These doorbells
are children of the managed partition nodes in the manager's device tree. These
doorbells are identified by an additional string in the compatible property:
"fsl,hv-state-change-doorbell", "fsl,hv-watchdog-expiration-doorbell", or
"fsl,hv-reset-request-doorbell".

Update the partition data structure (guest_t) to add three variables for
the doorbells that the managed partition rings for each of the three events.
These variables are only initialized for and used by managed partitions.

Create the send_doorbells() function, and update fh_partition_send_dbell() to
use it.

Note 1) This design does not conform to section 4.8.1 of the Topaz spec.
Instead of an "interrupts" property with three interrupt specifiers, we have
three doorbells per managed partition, and additional properties to help the
manager partition distinguish among them.

Note 2) A future version of this code will eliminate the need for adding nodes
to the DTS files.  Instead, the hypervisor will create all nodes automatically.
For now, however, this makes the code easier to implement and easier to
understand.

Note 3) This patch does not include any code to actually ring the doorbells
when the corresponding events occur.  That will be added later.

Signed-off-by: Timur Tabi <timur@freescale.com>