Handle ISI in hypervisor, for reflecting machine checks.
The VF bit only affects data accesses, not instruction fetches.  Thus,
we handle all ISI exceptions in the hypervisor to check for the VF bit.
Bug 2196.

ISI should not be as performance sensitive as DSI.

Also, when setting VF, store the original access bits.  Eventually,
this will be used for tlbsx/tlbre, and for reflecting an ISI rather
than a machine check if the guest did not request the type of access
that is being performed.

Signed-off-by: Scott Wood <scottwood@freescale.com>