hv: automatically create special doorbells
Have the hypervisor create the special doorbells automatically, instead of
requiring them to be in the DTS.

Rearrange init_guest_primary() so that the xxx_dbell_partition_init()
functions are called before process_partition_handles().
process_partition_handles() creates receive doorbell handles that do not have
endpoints, and recv_dbell_partition_init() considers that to be an error.

Factor out the code that attaches a doorbell to a receive handle into function
attach_receive_doorbell(), and update recv_dbell_partition_init() to call it.

Signed-off-by: Timur Tabi <timur@freescale.com>