partman: copy ELF images more intelligently
Currently, partman creates a giant local buffer and expands all ELF images into
the buffer before copying the entire buffer (holes and all) to the target
partition.  This does not work well for PT_LOAD segments that are far apart.
It also fails if the memory it copies to is not contiguous.  This can happen,
for example, if on PT_LOAD segment points to shared memory area that is not
adjacent to main memory.

Now, as we parse each PT_LOAD segment, it is copied to the target partition.

Signed-off-by: Timur Tabi <timur@freescale.com>