livetree: Add flatten_dev_tree().
This creates a flat device tree blob from a live tree.  I tested this
by turning the input fdt into a livetree and back again, succesfully.

Some changes to for_each_node to allow both pre- and post- actions,
to fix a bug on backtracking, and to use proper doxygen.

Signed-off-by: Scott Wood <scottwood@freescale.com>