hv: add I2C device register virtualization
Add support for virtualizing I2C device register access.  Each fsl-i2c node
in a guest's device tree has its device range registered with the
virtualization trap handler.  The page of the device is not mapped in the TLB.
When the guest attempts to access the page, a data storage execption is
generated.  If the address range is within the range of the device, then
the hypervisor performs the I/O operation and emulates the instruction that
caused the trap.  Control is returned to the guest, which is none the wiser.

Signed-off-by: Timur Tabi <timur@freescale.com>