Switch to livetrees.
The hardware tree is unflattened as soon as malloc() is up and
running, and the guest tree is flattened just before transfering control.

This will make it easier to hook extra information into nodes for
internal use (just add fields to the node struct), and makes node
copying/merging easier.

Some minor fixes and cleanups as well.

Signed-off-by: Scott Wood <scottwood@freescale.com>