hv: add support for virtualizing I2C device interrupts
Add the ability to virtualize the interrupts for any node in the guest device
tree.  The hypervisor will create a virtual interrupt to replace the hardware
interrupt specified in the node, and then call an internal ISR when the
hardware generates the interrupt.  If the ISR detects the interrupt came from
the device it's virtualizing, then the ISR clears the interrupt status in the
device and asserts the virtual interrupt.  The status register for the device
is also virtualized.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>