Do not disable waiting for gevent when returning to wait_for_gevent_loop.
Normally, the interrupted value of waiting_for_gevent is preserved, so
that it can be set to zero during the gevent so that a gevent cannot
interrupt a gevent if it enables critints (such as the gdb event).

However, wait_for_gevent() sets waiting_for_gevent (with critints
disabled, so there's no nesting) and returns into a loop rather than the
context it interrupted.  When this happens, we want to preserve the
new value of waiting_for_gevent rather than restore zero, which would
cause us to get stuck in waiting_for_gevent_loop without the ability
to execute gevents.

Signed-off-by: Scott Wood <scottwood@freescale.com>