hv: add dual-core support to watchdog unit test
Update the watchdog unit test so that it tests watchdog expiration on a
secondary core.  Previous, the test run only on the boot core.  Now it
runs the final reboot test on the second core, and it also shows the value
of the TSR register when the core starts.  This lets the tester check to
make sure that TSR[WRS] is equal to TCR[WRC] from the previous boot.

Signed-off-by: Timur Tabi <timur@freescale.com>