hv: change from "fsl,hv-pic-coreint" to "fsl,hv-pic-legacy"
Previously, the default mpic interface was legacy mode, and if the guest OS is
supposed to use coreint, its device tree must contain "fsl,hv-pic-coreint".
The new way is to assume coreint, and if the guest OS is supposed to use
legacy mode, the device tree must contain "fsl,hv-pic-legacy".

Signed-off-by: Timur Tabi <timur@freescale.com>