Don't map ranges if it's been delete-propped.
This has the side-effect of allowing the user to modify ranges
with node-update, and have the modifications be honored in map-ranges
and in rewriting the ranges property when reparenting.  It may be
useful, so I'm not going to go out of my way to break it.

It's probably not worth supporting in reg as well, though, as that
would require extra work.

Signed-off-by: Scott Wood <scottwood@freescale.com>