TLB cache: no-op global invalidates when cache is already clean
After a global invalidate, mark the TLB cache as clean.  Subsequent
tlbwe instructions will mark it dirty.  If another global invalidate
occurs while the TLB cache is still clean, treat it as a no-op.

This helps performance with software that makes bad assumptions about the
efficiency of global invalidates, and does not take care to avoid issuing
them unnecessarily.

Signed-off-by: Scott Wood <scottwood@freescale.com>