hv: add support for node-update-phandle configuration nodes
The node-update-phandle configuration subnode is used to set phandles for
guest nodes based on a hardware node.  Each property in node-update-phandle
points to another configuration node that references a node in the hardware
tree.  The phandle of that node is put into the guest node.  If that hardware
node does not have a phandle, then one is created.

Also fix a source code formatting error in merge_pre().

Signed-off-by: Timur Tabi <timur@freescale.com>