hv: fix handling of multiple phandles per property in node-update-phandle
If a configuration property in a node-update-phandle node specified more than
one phandle, the generate property in the guest device tree would only contain
the last phandle.

Also fix a panic if the phandle points to a configuration node that does not
create a node in the guest device tree (e.g. has no device= property).

Signed-off-by: Timur Tabi <timur@freescale.com>