hv: allow for skipping of automatic loading of images for debugging
If an image load table is defined, the hypervisor will always load the images
when the partition starts.  For debugging, this can be a problem because the
process can take a long time, and the user will usually want to overwrite those
images with his own anyway.  So we add a variable to the guest_t structure, so
if non-zero, the hypervisor will not load the images.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>