hv: obtain the CCSR base physical address from the hardware device tree
Look up the SOC node in the device tree and fetch the base physical address
of the CCSR address range.  Preferably, there will be a "ccsr" alias that
points to the SOC node.  If not, the hypervisor will look up the /soc node.

CCSRBAR_PA is changed from a macro to a variable.  CCSRBAR_SIZE is deleted
because it's never used.

Signed-off-by: Timur Tabi <timur@freescale.com>