Fix default allocation for CPC ways.
Transactions for a particular PMA should allocate in CPC, even if
"allocate-cpc-ways" property is not specified for it.

Also, we should not consider the already programmed values in the
CPCP[IAW]Rn registers. CPCP[IAW]R0 registers value should be set
to hardware reset values. This would ensure CPC to operate as a
non partitioned cache (all ways shared across all PMAs).

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>