added support for fsl,corenet-law and fsl,corenet-cf.
Patch for BugIDs 2639 and 2640. We get the law and ccf register information
from the hardware device tree, based on the above mentioned compatible strings.
We would need corresponding law and ccf device nodes in the hv-config node
of the config device tree. Added an additional check for pma initialization
for the case where CCM initialization fails.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>