hv: report error if node-update-phandle config property points to invalid node
The nodes referenced by properties in a node-update-phandle configuration node
most have valid endpoints.  That is, they must point to hardware nodes that have
a presence in the target guest device tree.

Add a check to make sure the referenced node does indeed have an endpoint, and
display a message if it doesn't.  Previously, the hypervisor would just panic.

Signed-off-by: Timur Tabi <timur@freescale.com>