CPC reconfiguration patch.
While configuring CPC partitioning following steps need to be followed (as per CPC block
guide 0.902) :
1. clear all bits of CPCPAR<n> (this prevents any new transactions from allocating
   into the CPC) using write operation
2. flush the CPC by setting CPCCSR0[CPCFL]
3. wait for hardware to clear CPCCSR0[CPCFL] using a polling loop
4. clear all lock bits by setting CPCCSR0[CPCLFC]
5. wait for hardware to clear CPCCSR0[CPCLFC] using a polling loop
6. disable the CPC by clearing CPCCSR0[CPCE]
7. reconfigure CPC as desired
8. enable CPC by setting CPCCSR0[CPCE]

The current CPC code does not follow these steps. This patch modifies the existing code to
follow these steps. We re-enable CPC once ccm_init is complete.

This patch has a dependency on simics model p4080-4.0.pre23, scheduled for release on February 23.
The current simics functional model does not support the auto-clearing functionality for CPCFL bit.
This results in a hypervisor hang while waiting for CPCFL bit to be cleared by the hardware.
This has been reported as a simics bug 12772.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>