Strip execute permissions from VF pages (bug 3007).
When TLB cache is enabled, the original exec permissions will be
remembered when doing tlbsx/tlbre.  There is no good way to do this
with TLB cache disabled.

The original MAS3 bits are not yet used to determine whether to reflect
a machine check or an ISI; the former is always reflected for VF pages.
This is a TODO.

Signed-off-by: Scott Wood <scottwood@freescale.com>