set_law: Properly set the high 4 bits of the physical address.
The >> 28 was simply wrong, and caused the high 4 bits to always be zero.
Plus, it was always setting the high-address register of LAW zero, rather
than indexing into the array.

Also, don't OR the new value into the old value.  That currently works
because the initial value of the LAW registers is zero, but it doesn't
make any sense in the general case.

Signed-off-by: Scott Wood <scottwood@freescale.com>