Support for read-only property in gpmas (BUG ID 3260)
Implement a boolean "read-only" property on guest physical memory areas.

When 'read-only' is specified any TLB mappings to that GPMA created by the by
the guest should be marked as user/supervisor read only.   The guest cannot
modify the user and supervisor write permission bits if this property is set--
writes to the UW/SW bits are ignored/masked-off.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>