do_merge_phandle: Change "config" to "src".
When I see "config node", I think of the node with the device property,
not any node in the config tree.  Change the name to clarify that
the node is the merge source, not the device config node.

Also remove the incorrect function name in the doxygen comment (doxygen
should provide the function name itself), remove a duplicate comment that
was copied rather than moved, and fix a comment typo.

Signed-off-by: Scott Wood <scottwood@freescale.com>