do_merge_phandle: Use guest_phandle.
This allows hypervisor-generated nodes to be referenced
in node-update-phandle (bug 2594).

Also, output both parent and node in error messages (the latter is
useless if the node is node-update-phandle itself, but can be helpful
if it is a child node (especially if it sits directly under
node-update-phandle)).  Previously, one test printed the node, and the
other the parent.

Note that phandles to node-update-phandle nodes (and children thereof)
are *not* supported reliably, because phandle connections must be
recorded before the node-update-phandle nodes are processed.  To have a
phandle to a node which is only referenced in node-update-phandle, the
user will need to create an empty node at the same path with node-update,
and refer to that.

Signed-off-by: Scott Wood <scottwood@freescale.com>