Implement device-ranges, and add a simple test case.
When a guest has a device-ranges, assigned devices will be placed
into a container that has a ranges property equal to the contents
of device-ranges.

While this does require the guest to support 64-bit intermediate values
in device tree parsing (which it will if it reuses our device tree code,
which we should put in libos -- not sure what LWE uses), it avoids
problems with translating ranges that overlap device-ranges entries.

Signed-off-by: Scott Wood <scottwood@freescale.com>