pamu: fixes and cleanup
Fix error messages to use the right loglevel, and not call errors warnings.

Don't error out if one subwindow is bad -- continue with the rest.

Do error out if the subwindow count is invalid -- don't treat it as
if no subwindows were specified.

Complain when a subwindow has a bad address; previously it was just ignored.

If subwindow size is missing, default to the maximal size -- but warn.

Set the SPAACE valid bit at the end, just in case.

Reduce the subwindow setup to one loop rather than two, and set up
the PPAACE there rather than in the parent to avoid having to pass
a bunch of data back.

Do not reference anything in the device type with variable-sized
types such as phys_addr_t.

Add some extra checking of window sizes and alignments.

Handle a primary window base address >4GiB.

Handle subwindow nodes that are not in ascending order.

Signed-off-by: Scott Wood <scottwood@freescale.com>