Flexible IMA specification (Bug 3509).
A partition can now specify virtual address, physical address, and
size of the boot core's IMA (at the likely expense of ePAPR compliance,
though that's still a FIXME in the default case).

Tested manually, but a thorough automated test is a bit tricky if we want
to test any interesting values -- they won't work with our normal test
environment that expects to start at gphys zero.  Even if the linker
script is tweaked, it's hard to verify exactly what virtual address
we originally came in on.

Signed-off-by: Scott Wood <scottwood@freescale.com>