Fix gdebug test hang
Bugid 3598.

With the new toolchain version gcc-4.3.74-eglibc-2.8.74-2, the compiler simply
removed the call (bl) to the function "hwbreaktestfunc" and inlined it in the
caller code. IAC breakpoint was specifically set for this function.Now as this
address wasn't getting hit IAC exception wasn't generated, thus the test got hung.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Signed-off-by: Ashish Kalra <Ashish.Kalra@freescale.com>
Signed-off-by: Scott Wood <scottwood@freescale.com>