bcmux: Fix mismatched error code in __mux_send_data.
When the queue was full, it would usually return 0, but sometimes -1.
The callers were expecting -1, so pull-mode data transmission did not
generally work.  This can be seen by printing a large device tree
through the hypervisor console attached to a mux; it would halt
partway through until a key is pressed to wake the thread.

It now always returns -1 when the queue is full.

Signed-off-by: Scott Wood <scottwood@freescale.com>