Avoid redundant PID invalidations.
Linux has a habit of issuing several full-PID invalidations back-to-back,
due to inefficient communication between generic and arch code.  While this
should be fixed, for now we avoid invalidating a PID if the last thing
we did to the TLB was to invalidate that same PID.

We already do this for full-LPID invalidations, as Linux used to generate
those instead.

Signed-off-by: Scott Wood <scottwood@freescale.com>