Descriptive error message for partman.
When partman get's a error code due to a failing hcall, user should
understand the reson for the failure. This patch maps the error code
to descriptive error message.

Earlier the error print could be suppresed by the "quiet" option,
which should not be the case. Now, this error message would be displayed
irrespective of the quiet option. Other error messages can still be suppressed
using the quiet option.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>