PAMU: Don't fail if we can't set up a LAW.
If we don't have access to the LAWs, then we won't have set up a LAW for
the hypervisor's PMA, so this isn't fatal (as long as it wasn't due to
running out of LAWs).  Worse, pamu_hw_init() has already been called, so
it's too late to abort and free the memory that the PAMU is looking at.

Signed-off-by: Scott Wood <scottwood@freescale.com>