Don't disable CPCs while setting the partitioning registers.
CPCs should not be disabled once DDR controllers are enabled and configured to use
memory interleaving (BugID 4698).

When hypervisor comes up it may require to partition the CPC ways as per partition
requirements. Earlier we were disabling the CPCs (as directed by the block guide)
while configuring the partitioning registers, but as per CPC errata8 this can cause
problems in an interleaved configuration. The fix for this errata is, to not disable
CPC during hypervisor initialization.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>