Use xtel.sh instead of manual mux_server/xterm.
Added a local version of xtel.sh free of sim.git maintenance issues,
without the geometry stuff (let the window manager choose).

xtel.sh is now invoked with setsid, which is the only way I could find to
prevent ctrl-C in simics from killing the mux server.  I think executing
from make introduces some weirdness here.  The downside to this is that
if something goes wrong, the mux server or the xterms may not get killed
when simics exits, because this also stops the hangup signal.  When
things are working, the mux server will exit when its network connection
goes away.

Signed-off-by: Scott Wood <scottwood@freescale.com>