test makefiles: use double-colon rules for run
This allows the individual makefile to specify supplemental run actions
(such as xtel, though we eventually may want to make that part of the
common makefile) while still using the common makefile, without causing a
duplicate rule message.

Note that in order for the supplemental action to come before simics, the
rule must be declared before the common Makefile is included.

Signed-off-by: Scott Wood <scottwood@freescale.com>