managed partitions: get the node name from the config node.
This is how other nodes work (assigned devices, byte channels, etc), and
it works well for ensuring that the guest device node has a valid, unique
name.

Previously it took the name from the partition's label, which may not be
a valid node name, and is inconsistent with other node types.

Existing tests are changed to have cfgnode name == label, to eliminate
any chance of breakage.

Signed-off-by: Scott Wood <scottwood@freescale.com>