Enter nap from the idle loop.
This expands the use of nap for all cases where the hv is idle, except on
the boot core.  In theory this is selected based on whether gcpu->napping
is set, but currently there is nothing that would cause a non-boot core to
go idle with gcpu->napping clear, and there would be added complexity in
making the decision dynamically.

gevents now wake up a napping core, and tlbivax will now properly skip a
napping core.

Issues that were raised with the cache flushing will be addressed in
a future patch.

Signed-off-by: Scott Wood <scottwood@freescale.com>