nap: misc fixes
Clear d-cache locks before flushing (working around erratum CPU20 in the
process).

Be sure to use a private region for flushing, as snoops can disturb the
process.  Switch to a dcbz-based flush with flush assist, to reduce the
amount of displacement space needed on each core, and speed up the
process.

Replace magic constants with symbolic names for some of the cache bits.

Set L2CSR0[L2DO,L2IO] separately from L2CSR0[L2FL].  This avoids a
problem whereby L2FL will sometimes not clear itself.

Avoid napping or otherwise touching the rpcm registers when we don't
have them.

Check for reflecting a trap to an idle thread after checking that it's
not really a hypervisor crash.

Signed-off-by: Scott Wood <scottwood@freescale.com>