emulate the fixed interval timer (FIT)
Instead of allowing the FIT interrupts to be routed directly to the guest,
route them to the hypervisor, which will reflect them to the guest.  Also add
a mechanism whereby the hardware FIT can be set to a higher frequency than
what the guest programmed it for, and the hypervisor will reflect every 'n'
interrupts instead.

In gcpu_t, rename 'tsr' to 'watchdog_tsr', since this variable is only used to
keep track of the TSR register during a watchdog reset.  Rename 'timer_flags' to
'gtcr', and introduce 'gtsr' for virtualizing the guest TSR register.

Rename TIMER_FLAGS_MASK to GTCR_FLAGS_MASK, and introduce GTSR_FLAGS_MASK.

Remove the GCPU_PEND_FIT and GCPU_PEND_TCR_FIE flags.  We don't disable the
hardware FIT interrupt any more, because we only reflect the interrupt to the
guest when it's expecting one, so there's no need to toggle it on and off.
We also don't need a FIT pending flag for guest_doorbell(), because this
function now uses the guest SPRs to determine what to do.  This means that
we also don't need run_deferred_fit() or enable_tcr_fie().

Also update the FIT unit test.

Signed-off-by: Timur Tabi <timur@freescale.com>