Add claimable device support.
To support partition failover, devices can be assigned to multiple
partitions in a manner where only one partition has active ownership
at a time.  After that partition is stopped, a standby owner of the device
can claim the device and become the new active owner.

On a device claim, interrupts and DMA errors are rerouted.

TODO: Migrate existing access violations that are sitting in the
previous active's queue.

The unit test will currently fail on hardware because of an MPIC issue
that is being investigated.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>