Clear coreint interrupts on partition stop by temporarily clearing EXTGS.
Neither masking an interrupt nor doing an EOI loop are sufficient to clear
a coreint interrupt that has already been pushed to the core, even if
it has not been taken yet.

We temporarily turn off EXTGS and take the interrupts in the hypervisor
(doing nothing but EOI in the handler) to clear these interrupts.

This clearing of interrupts must happen after the interrupt has been
masked, but before any other reconfiguration takes place, so handle
reset is split into two parts, one before per-cpu cleanup and one after.

A new state is introduced to synchronize the per-cpu cleanup, but we avoid
letting it be part of the public API by having an internal state range
in which all states map to guest_stopping.

The failover test now checks for spurious even with coreint, as that's
how simics handles the withdrawn interrupt on coreint.

Signed-off-by: Scott Wood <scottwood@freescale.com>