fix emulation of TCR and TSR register bits
Some of the code that handled the emulation of various bits of the TCR and TSR
registers was sloppy.  Technically, it all worked, but it made assumptions that
may have never been true or were true only until recently (e.g. the watchdog
emulation feature).

Delete GTSR_FLAGS_MASK, since it wasn't being used.  All bits of TSR are now
emulated for the guest.

Replace GTCR_FLAGS_MASK with GCPU_TCR_HW_BITS, which designates all of the TCR
bits that the guest can modify in the hardare TCR.

Update the guest initialization functions so that they set/clear the correct
bits in TCR and TSR.

Fix set_tcr() so that it only modifies the TCR bits that should be modified.

Finally, update the 'timer' unit test and add some documentation to get_tsr().

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Scott Wood <scottwood@freescale.com>