Implement zero-pma.
PMAs with this property will be zeroed on hv boot when the PMA is processed.

PMAs will only be processed on boot if at least one partition (or the hv
itself) references it with a gpma.  Non-referenced PMAs are not a supported
configuration.

The failover test is updated to use zero-pma instead of an "initialize-me"
device tree property.

Signed-off-by: Scott Wood <scottwood@freescale.com>