add hypervisor watchdog support
Add support for using the hardware watchdog timer to the hypervisor.  If the
watchdog-enable property is found, its value is used to enable and configure
the watchdog.  If critical interrupts are disabled for too long, the system
will reset.

Currently, watchdog runs only on core 0.  Support for secondary cores will be
added later.

Signed-off-by: Timur Tabi <timur@freescale.com>