Modified error_manager_guest global storage, split the unit tester, sent doorbells to error manager
Restored the global storage of the error manager guest to a guest_t *. There could be accessors to it
implemented later.
Events in error queue are stored even if there is no error manager active but now a critical interrupt
is dispatched in case of past events in the queue.
Split the unit tester into two test suites selectable with a suffix on the partition name: e.g. part1-A.
There are two test suites now: A and B.
Applied two patches from Scott:
- error-mgr.c::test_claim_error_manager(): "The gcpu only determines which core gets the interrupt.
There is no restriction on which gcpu makes hcalls related to the error queue, including claim."
- devtree.c::claim_error_manager(): "Now that the race that the comment in this code was worried about is
fixed, stop being paranoid about whether "prev" is really the correct
previous owner, or still stopped.  Also don't be paranoid about whether
the new owner struct's guest field is correct."

Signed-off-by: Cristian CACIULOIU <ccaciul1@freescale.com>