Use same liodn offset for all QMan portals within a partition.
This is required to work around a design issue in the CAAM.

Because all portals within a partition must now have the same DMA
config for its device children (the portal's own LIODNs can still
be different), the device children of a qman portal are moved out
into a portal-devices node under the partition config node.

Old-style config trees will still work, with warnings, and the
config of the first portal in the partition will be used.

Signed-off-by: Scott Wood <scottwood@freescale.com>