vpic_alloc_handle: take standby argument
Previously, vpic_alloc_handle always told vmpic_alloc_handle that it is not
standby, which results in an interrupt handle that the standby guest can
unmask.

If there happens to be no active interrupt due to a misconfiguration, and
the standby guest unmasks such an error interrupt, it can cause the interrupt
to be unmasked with no handler, which can lead to an hv assert.

Signed-off-by: Scott Wood <scottwood@freescale.com>