TLB related fixes and isync after MAS writes
Fixed several locations where isync was not placed after the last MAS
write and a TLB management instruction.
Added isync after writing to SPR_MMUCSR0.
Fixed a bug related to searches in TLBC of conflicting pages with an
tentative write in TLB1, where the number of 4K segments to be searched
was too small (maximum 12).
Fixed a bug found by Scott Wood related to incorrect search in TLB0 for
the case of HV without TLBC, where TLBSX did not receive an updated EPN.
Added a test case in MMU for conflicting TLB settings.

Signed-off-by: Cristian CACIULOIU <ccaciul1@freescale.com>