Check for full output queue even if the current byte channel is currently selected.
In pulll mode, if the current byte channel is currently selected for transmission,
the test for output queue size is no longer made. This may end up in a situation
when the pull mechanism is off, but one channel queue is full. In this case this
channel cannot send more data until the pull mechanism is again activated.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>