Make clearing PAMU[AVICS] revision dependent.
PAMU 3 erratum workaround suggests writting the PICS[AVICS] bit twice.
For rev2 silicon, this behavior was fixed, but PAMU-A001 says that
software writting 1 to PAMU registers would have the expected result
(to clear the bit) only if the bit to be clearead is 1, otherwise it
just sets the value of register to 1. In this case clearing PAMU[AVICS]
should be handled differently depending on the silicon revision.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>