Fixed the cache-A unit tester.
The cache-A unit tester uses an algorithm to evict the L1
cache. The algorithm is based on accessing the cache for a
certain number of times. The test was accessing the L1 cache
with store type operations, however the cache behaves differently
on store miss situations depending on the cache configuration
(write shadow or not). Changed the behavior and used load type
operations to access the cache.

Signed-off-by: Diana CRACIUN <Diana.Craciun@freescale.com>