Initialize breakpoint address to -1
The breakpoint table is initialized to 0x0's when allocated, however
0x0 is a valid address. The logic in locate_breakpoint() goes by
the address, so we end-up with a false match for address 0x0 and
assume that we have a breakpoint at 0x0 - that is not right. Hence,
we now initialize/re-initialize a breakpoint with an invalid address
of -1 upon allocation/"deletion" respectively.

Signed-off-by: Anmol P. Paralkar <anmol@freescale.com>
Signed-off-by: Laurentiu TUDOR <Laurentiu.Tudor@freescale.com>